tonsbergtur/requirements.txt
wheel
Django>4.0.0
django-ckeditor
pillow
django-sslserver
djangorestframework
geopy
python-Levenshtein